requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0
mediapipe>=0.10.0
transformers>=4.35.0,<5.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
tqdm>=4.65.0
